requirements.txt
mcp>=1.26.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.20.0
